[fix] Priority Queue not working properly
I was using a priority queue to order my nodes for A*. I had customised the comparison operator in the Node class yes, but it wasn't a queue filled with Nodes, it was a priority queue filled with POINTERS of Nodes. So the PQ was outputing nodes based on their memory adress, thus giving random results.
So I changed the code to add a custom Comparator to the priority queue that actually compares something and now everything works properly.